src/test/java/Othello/BoardTest.java
package Othello;

import static org.junit.Assert.*;

import org.junit.Before;
import org.junit.Test;

public class BoardTest {
	Board board;
	Disk[][] auxBoard;

	@Before
	public void setUp() throws Exception {
		board = new Board();
		auxBoard = board.getGameBoard();
	}

	@Test
	public void BoardConstructorTest() {
		
		for(int i=0; i<auxBoard.length; i++) {
			for(int j=0; j<auxBoard.length; j++) {
				if((i==3 && j==3) || (i==4 && j==4)) {
					assertEquals(Color.White, auxBoard[i][j].getColor());
				}
				else if ((i==3 && j==4) || (i==4 && j==3)) {
					assertEquals(Color.Black, auxBoard[i][j].getColor());
					}
				else {
					assertEquals(null, auxBoard[i][j]);
					}
			}
		}
	}
	
	@Test
	public void isFullTest() {
		board.setTotalWhites(25);
		board.setTotalBlacks(32);
		boolean res_25W_32B = board.isFull();
		assertFalse(res_25W_32B);
		
		board.setTotalWhites(32);
		board.setTotalBlacks(32);
		boolean res_32W_32B = board.isFull();
		assertTrue(res_32W_32B);
	}
	
	@Test
	public void updateScoreTest() {
		board.updateScore();
		assertEquals(2, board.getTotalBlacks());
		assertEquals(2, board.getTotalWhites());
		
	}
	
	@Test
	public void outOfLimitsTest() {

		boolean res_inside = board.outOfLimits(2, 3);
		assertTrue(res_inside);
		
		boolean res_x_out_y_inside = board.outOfLimits(9, 4);
		assertFalse(res_x_out_y_inside);
		
		boolean res_x_inside_y_out = board.outOfLimits(2, 10);
		assertFalse(res_x_out_y_inside);
		
		boolean res_x_out_y_out = board.outOfLimits(9, 9);
		assertFalse(res_x_out_y_out);
		
	}
	
}
		
